// src/app/models/common-listing.ts
export class CommonListing {
    value: string | number;
    viewValue: string;
}
export class CommonList {
    id: string | number;
    name: string;
}

export class HeaderElement {
    label: string | '';
    value: string | '';
}

export class CommonListingClass {
    value: string | number;
    viewValue: string;
    class: string;
}